# dispatcher/backend/docs/openapi_v1.yaml
openapi: 3.0.0
servers:
  - description: production server
    url: https://api.farm.openzim.org/v1
info:
  description: |
    Zimfarm is openZIM's automated & distributed ZIM files creation engine.
    openZIM periodically publishes updated ZIM files for a large variety of content.

    This API provides public access to the scheduled ZIM files to create as well as tracking of its tasks.

    This API is also used by _workers_ to fetch, process and update those tasks.
  version: 1.0.0
  title: Zimfarm API
  contact:
    email: contact+zimfarm@kiwix.org
    url: https://kiwix.org
  license:
    name: GNU General Public License v3.0
    url: https://www.gnu.org/licenses/gpl-3.0.en.html
...
paths:
...
  /schedules/:
    get:
      tags:
      - public
      summary: list schedules
      operationId: listSchedules
      description: Query the Schedules database
      parameters:
      - $ref: '#/components/parameters/SkipParameter'
      - $ref: '#/components/parameters/LimitParameter'
      - in: query
        name: category
        description: Categories to filter results by (union of, if several)
        required: false
        schema:
          type: array
          items:
            type: string
          example:
            - other
            - wikipedia
      - in: query
        name: tag
        description: Tags to filter results by (intersection of, if several)
        required: false
        schema:
          $ref: '#/components/schemas/ScheduleTags'
      - in: query
        name: lang
        description: Language code to filter results by (union of, if several)
        required: false
        schema:
          type: array
          items:
            type: string
            format: ISO 639-1
          example:
            - fr
            - bm
      - in: query
        name: name
        description: Name to filter results by (direct match or regex)
        required: false
        schema:
          type: string
        example: wikipedia_.+_football
      responses:
        200:
          description: List of Schedules matching criteria
          content:
            application/json:
              schema:
                type: object
                properties:
                  meta:
                    $ref: '#/components/schemas/MetaProperty'
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/ScheduleItem'
        400:
          description: Bad Request (invalid input)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InputError'
    post:
      tags:
      - restricted
      summary: create schedule
      operationId: addSchedules
      description: Create a new Schedule
      security:
        - token: []
        - oauth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Schedule'
      responses:
        201:
          description: Schedule Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  _id:
                    $ref: '#/components/schemas/MongoId'
        401:
          description: Unauthorized
        400:
          description: Bad Request (invalid input)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InputError'
...
  /schedules/{scheduleName}:
    get:
      tags:
      - public
      summary: list schedules
      operationId: getSchedule
      description: Retrieve a Schedule's details
      parameters:
        - $ref: '#/components/parameters/ScheduleNameParameter'
      responses:
        200:
          description: Schedule details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Schedule'
        400:
          description: Bad Request (invalid input)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InputError'
        404:
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    patch:
      tags:
      - restricted
      summary: update schedule
      operationId: editSchedule
      description: Update a Schedule's details
      security:
        - token: []
        - oauth: []
      parameters:
        - $ref: '#/components/parameters/ScheduleNameParameter'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                name:
                  $ref: '#/components/schemas/ScheduleName'
                language:
                  $ref: '#/components/schemas/Language'
                category:
                  $ref: '#/components/schemas/ScheduleCategory'
                tags:
                  $ref: '#/components/schemas/ScheduleTags'
                enabled:
                  type: boolean
                config:
                  $ref: '#/components/schemas/Config'
      responses:
        200:
          description: Updated
        400:
          description: Bad Request (invalid input)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InputError'
        401:
          description: Unauthorized
        404:
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      tags:
      - restricted
      summary: delete schedule
      operationId: deleteSchedule
      description: Remove a Schedule
      security:
        - token: []
        - oauth: []
      parameters:
        - $ref: '#/components/parameters/ScheduleNameParameter'
      responses:
        200:
          description: Schedule Deleted
        400:
          description: Bad Request (invalid input)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InputError'
        401:
          description: Unauthorized
        404:
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /schedules/{scheduleName}/clone:
    post:
      tags:
      - restricted
      summary: clone schedule
      operationId: cloneSchedule
      description: Clone a Schedule into another
      security:
        - token: []
        - oauth: []
      parameters:
        - $ref: '#/components/parameters/ScheduleNameParameter'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                name:
                  $ref: '#/components/schemas/ScheduleName'
      responses:
        201:
          description: Schedule cloned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Schedule'
        400:
          description: Bad Request (invalid input)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InputError'
        401:
          description: Unauthorized
        404:
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    token:
      type: apiKey
      description: API Token
      name: token
      in: header
    oauth:
      type: oauth2
      description: OAuth2
      flows:
        password:
          tokenUrl: https://api.farm.openzim.org/v1/auth/oauth2
          refreshUrl: https://api.farm.openzim.org/v1/auth/oauth2
  parameters:
    SkipParameter:
      in: query
      name: skip
      description: number of records to skip for pagination (yay)
      required: false
      schema:
        $ref: '#/components/schemas/SkipProperty'
    LimitParameter:
      in: query
      name: limit
      description: maximum number of records to return
      required: false
      schema:
        $ref: '#/components/schemas/LimitProperty'
    ScheduleNameParameter:
      in: path
      required: true
      name: scheduleName
      schema:
        $ref: '#/components/schemas/ScheduleName'
...
  schemas:
    Language:
      type: object
      required:
      - code
      - name_en
      - name_native
      properties:
        code:
          type: string
          format: ISO 639-1 language code
          example: fr
        name_en:
          type: string
          example: French
        name_native:
          type: string
          example: Français
...
    ScheduleItem:
      type: object
      required:
        - config
        - language
        - name
        - category
      properties:
        config:
          type: object
          required:
            - task_name
          properties:
            task_name:
              type: string
        category:
          $ref: '#/components/schemas/ScheduleCategory'
        language:
          $ref: '#/components/schemas/Language'
        name:
          type: string
        most_recent_task:
          type: object
          required:
            - _id
            - status
            - updated_at
          properties:
            _id:
              $ref: '#/components/schemas/MongoId'
            status:
              $ref: '#/components/schemas/TaskStatus'
            updated_at:
              type: string
              format: date-time
    ScheduleCategory:
      type: string
      example: wikipedia
      enum:
        - gutenberg
        - other
        - phet
        - psiram
        - stack_exchange
        - ted
        - vikidia
        - wikibooks
        - wikinews
        - wikipedia
        - wikiquote
        - wikisource
        - wikispecies
        - wikiversity
        - wikivoyage
        - wiktionary
    Schedule:
      type: object
      additionalProperties: false
      required:
        - category
        - config
        - enabled
        - language
        - name
        - tags
      properties:
        category:
          $ref: '#/components/schemas/ScheduleCategory'
        config:
          $ref: '#/components/schemas/Config'
        enabled:
          type: boolean
          example: true
        language:
          $ref: '#/components/schemas/Language'
        name:
          $ref: '#/components/schemas/ScheduleName'
        most_recent_task:
          type: object
          required:
            - _id
            - status
            - updated_at
          properties:
            _id:
              $ref: '#/components/schemas/MongoId'
            status:
              $ref: '#/components/schemas/TaskStatus'
            updated_at:
              type: string
              format: date-time
        tags:
          $ref: '#/components/schemas/ScheduleTags'
    ScheduleName:
      type: string
      example: wikipedia_fr_all
    Config:
      type: object
      required:
        - flags
        - image
        - resources
        - task_name
        - warehouse_path
      properties:
        flags:
          type: object
        image:
          $ref: '#/components/schemas/ConfigImage'
        resources:
          $ref: '#/components/schemas/ConfigResources'
        task_name:
          type: string
          example: mwoffliner
        warehouse_path:
          type: string
          example: /wikipedia
    ConfigImage:
      type: object
      required:
        - name
        - tag
      properties:
        name:
          type: string
          example: openzim/mwoffliner
        tag:
          type: string
          example: latest
    ConfigResources:
      type: object
      required:
        - cpu
        - memory
        - disk
      properties:
        cpu:
          $ref: '#/components/schemas/Cpu'
        memory:
          $ref: '#/components/schemas/Memory'
        disk:
          $ref: '#/components/schemas/Disk'
    Cpu:
      type: integer
      format: int32
      example: 3
      description: number of CPU cores required to run it
    Memory:
      type: integer
      format: int32
      example: 1024
      description: RAM (in bytes) required to run it
    Disk:
      type: integer
      format: int32
      description: Disk space (in bytes) required to run it
...
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          example: Not Found
    InputError:
      type: object
      properties:
        message:
          oneOf:
            - type: object
              example: {"schedule_name": "required field"}
            - type: string
              example: Invalid signature format (not base64)
...
    MetaProperty:
      type: object
      properties:
        count:
          type: integer
          example: 1
        limit:
          type: integer
          example: 10
        skip:
          type: integer
          example: 0
    LimitProperty:
      type: integer
      format: int32
      minimum: 0
      example: 10
      default: 10
    SkipProperty:
      type: integer
      format: int32
      minimum: 0
      example: 0
      default: 0
    MongoId:
      type: string
      format: uuid
      example: 5dbc599b3b61e21c362703e4
...
    ScheduleTags:
      type: array
      items:
        type: string
      example:
        - selection
        - full
        - top
    TaskStatus:
      type: string
      example: canceled
      enum:
        - started
        - scraper_started
        - scraper_completed
        - scraper_killed
        - failed
        - cancel_requested
        - canceled
        - succeeded
        - created_file
        - uploaded_file
...